requirements.txt
Django==1.10.4
django-extensions==1.7.5
django-graphos==0.3.10
Werkzeug==0.11.11
gunicorn==19.6.0
-e git+https://github.com/nkunihiko/django-bootstrap3-datetimepicker.git#egg=django-bootstrap3-datetimepicker
